requirements.txt
# Core dependencies
fastapi==0.103.2
uvicorn==0.22.0
pydantic==1.10.7
httpx==0.24.1
paho-mqtt==2.1.0
redis==4.5.5
PyJWT==2.7.0
python-dotenv==1.0.0

# Async support
aiohttp==3.8.4

# Testing


# Development
black==23.3.0
flake8==6.0.0
isort==5.12.0

pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-xdist==3.3.1  # For parallel execution
